add mtk swconfig

+++ diy.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 VERSION="V3.3.0-6"
 A=0
+B=0
 [ -n "$OP_TARGET" ] || OP_TARGET="X64"
 case "$OP_TARGET" in
 	X64)path="X64";;
@@ -9,22 +10,22 @@ case "$OP_TARGET" in
 	R7800)path="R7800";;
 	N1)path="N1";;
 	RPI-4)path="RPI-4";;
-	K2P-16M)path="K2P";A=1;;
-	K2P-32M)path="K2P";A=2;;
-	MI-AC2100)path="MI-AC2100";;
-	REDMI-AC2100)path="REDMI-AC2100";;
-	R2S)path="R2S";A=3;;
-	R4S)path="R4S";A=3;;
-	R1-PLUS)path="R1-PLUS";A=3;;
+	K2P-16M)path="K2P";A=1;B=1;;
+	K2P-32M)path="K2P";A=1;;
+	MI-AC2100)path="MI-AC2100";A=1;;
+	REDMI-AC2100)path="REDMI-AC2100";A=1;;
+	R2S)path="R2S";A=2;;
+	R4S)path="R4S";A=2;;
+	R1-PLUS)path="R1-PLUS";A=2;;
 	*)echo "No adaptation target!";exit 1;;
 esac
 cp -r target/$path/. Small_5
+[ $B = 1 ] && rm -f Small_5/Patch-K2P-32M.patch
 
 if [ $A = 1 ];then
-	rm -f Small_5/Patch-K2P-32M.patch
+	rm -rf openwrt/target/linux/ramips
+	cp -r target/ramips/. Small_5
 elif [ $A = 2 ];then
-	rm -f Small_5/Patch-K2P-16M.patch
-elif [ $A = 3 ];then
 	rm -rf openwrt/package/boot/uboot-rockchip openwrt/package/kernel/linux/modules/video.mk openwrt/target/linux/rockchip
 	cp -r target/rockchip/. Small_5
 fi
